// routes/logic/classes/interfaces/CompaniesInterface.ts
export interface CompaniesInterface {
    company : number;
    description : string;
    active : number;
    intercompany: number;
    intercompany_connection: number;
}

export interface CompaniesDataInterface {
    COMPANY: number;
    DESCRIPTION: string;
    ACTIVE: number;
    INTERCOMPANY: number;
    INTERCOMPANY_CONNECTION: number;
}